# Use comprehension instead of map for AbstractArray filter