It now works with `NullableArray` and appears to be faster on regular `Array` and especially on `SparseMatrixCSC`.