@JeffBezanson 2017-04-17
Won't this cause memory blowup for sparse arrays?