@iamed2 2017-04-17
At what size do you think that might happen? While testing with `sprand(1000, 1000, 0.2)` I found memory use was halved.

Old: 
```julia
julia> srand(0);

julia> s = sprand(1000, 1000, 0.2);

julia> @benchmark filter(x->x<=0.5, $s)
BenchmarkTools.Trial:
  memory estimate:  11.64 MiB
  allocs estimate:  12
  --------------
  minimum time:     51.270 ms (0.00% GC)
  median time:      55.655 ms (0.00% GC)
  mean time:        56.322 ms (2.80% GC)
  maximum time:     119.802 ms (48.58% GC)
  --------------
  samples:          89
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
```

New:
```julia
julia> srand(0);

julia> s = sprand(1000, 1000, 0.2);

julia> @benchmark filter(x->x<=0.5, $s)
BenchmarkTools.Trial:
  memory estimate:  4.01 MiB
  allocs estimate:  9
  --------------
  minimum time:     26.978 ms (0.00% GC)
  median time:      28.984 ms (0.00% GC)
  mean time:        29.586 ms (0.63% GC)
  maximum time:     34.852 ms (0.00% GC)
  --------------
  samples:          169
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
```