@JeffBezanson 2017-04-17
It depends on the predicate. For a predicate that's more often false, `map` is better since it keeps things sparse, but if the predicate is often true then a dense `Bool` array is more compact. So I guess there's no one optimal choice.