@nalimilan 2017-04-17
The fact that the `map` approach doesn't work with `NullableArray` is an unfortunate consequence of automatic lifting, and shouldn't be taken as an indication of what's the best behavior for generic `AbstractArray`. We should add a custom method for `NullableArray` instead if that's the only/main reason for this change.