@JeffBezanson 2017-04-17
Actually, you have a point that the current `filter` has a worse worst case, which is representing a dense bool array as a sparse array (since the indexes are bigger than the data elements). So it seems reasonable to go with this. In some cases using a `BitArray` is even better:

```
julia> filtbitarray(f, As::AbstractArray) = As[BitArray(f(x) for x in As)];

julia> @time filter(x->x>0.5, s);
  0.332583 seconds (13.67 k allocations: 3.337 MiB)

julia> @time filt21423(x->x>0.5, s);
  0.330177 seconds (10.15 k allocations: 16.982 MiB)

julia> @time filtbitarray(x->x>0.5, s);
  0.353954 seconds (22.77 k allocations: 4.276 MiB)

julia> @time filter(x->x<0.5, s);
  1.006539 seconds (13.66 k allocations: 140.518 MiB, 1.20% gc time)

julia> @time filt21423(x->x<0.5, s);
  0.424649 seconds (10.14 k allocations: 18.209 MiB, 14.30% gc time)

julia> @time filtbitarray(x->x<0.5, s);
  0.370472 seconds (22.77 k allocations: 5.503 MiB)
```