@KristofferC 2017-04-17
Try something like a 500 000 x 500 000 matrix with ~ 10 numbers per column?