@nanosoldier 2017-04-18
[Your benchmark job](https://github.com/JuliaLang/julia/pull/21423#issuecomment-294595039) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/f3c288ba7387618401be3216a8366e24d7136d16/2ebc808_vs_2e89bc1/report.md). cc @jrevels