@ararslan 2017-07-26
Sorry, seems this fell by the wayside. Let's recheck Nanosoldier since it's been so long.

@nanosoldier `runbenchmarks(ALL, vs=":master")`