@tkelman 2017-07-26
For sparse arrays it might make sense to check for the result of the predicate called on zero, and branch to different behaviors whether it's true vs false.