@nanosoldier 2017-07-26
[Your benchmark job](https://github.com/JuliaLang/julia/pull/21423#issuecomment-318131362) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/400b395e9f6a130f3ca4b6298365417d74079e03/b18f5b8_vs_0a37b3d/report.md). cc @ararslan