@ViralBShah 2018-12-17
The right thing to do is what @tkelman suggested. I am closing this for now.